File upload without being logged in
A user can now upload files without being logged in. Of course, this
does not allow anyone to view the file again later, and it will not be
added to a user's profile.
The uploaded files will not be renamed and moved into an own folder, so
they can be easily removed with a cron job.